requirements.txt
python-dateutil>=2.8.0
transformers>=4.30.0
torch>=2.0.0
accelerate>=0.20.0